fix(core-004-type-generation): Fix BigInt compatibility and bindings export path
- Changed Attachment.size_bytes from i64 to i32 for TypeScript compatibility
- Fixed bindings export path in all Tauri apps (was ../../../packages, now packages)
- Domain types now correctly exported to TypeScript bindings
- All 14 enums and 15+ structs now appear in generated .ts files
- Fixed ESLint warning for unused TAURI_CHANNEL import

+++ backend/crates/altair-db/src/schema/shared.rs
@@ -64,9 +64,9 @@ pub struct Attachment {
     pub id: Option<Thing>,
     pub filename: String,
     pub mime_type: String,
-    pub size_bytes: i64,
+    pub size_bytes: i32, // Changed from i64 for TypeScript compatibility (max ~2GB)
     pub storage_key: String, // S3 key
-    pub checksum: String,    // SHA-256
+    pub checksum: String, // SHA-256
     pub media_type: MediaType,
     pub duration: Option<i32>, // seconds, for audio/video
     pub thumbnail_key: Option<String>,
